requirements.txt
streamlit==1.26.0
azure.cognitiveservices.speech 
pyaudio
wave
openai
azure.ai.translation.text 
azure.core 
